Separate table reference and helpers from `models.core`
