+++ database_client/result_formatter.py
@@ -10,11 +10,11 @@ from database_client.constants import (
 )
 from database_client.enums import LocationType
 from database_client.models.core import (
-    SQL_ALCHEMY_TABLE_REFERENCE,
     Agency,
     DataSourceExpanded,
     LocationExpanded,
 )
+from database_client.models.table_reference import SQL_ALCHEMY_TABLE_REFERENCE
 from database_client.subquery_logic import SubqueryParameters
 from utilities.common import format_arrays
