+++ database_client/subquery_logic.py
@@ -5,7 +5,7 @@ from pydantic import BaseModel
 from sqlalchemy.orm import joinedload
 from sqlalchemy.sql.base import ExecutableOption
 
-from database_client.models.core import convert_to_column_reference
+from database_client.models.helpers import convert_to_column_reference
 from middleware.enums import Relations
